Clean up code smells, grammar, code inspections etc.

+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/container/matroska/format/MatroskaEbmlReader.java
@@ -82,7 +82,7 @@ public class MatroskaEbmlReader {
   }
 
   private static long applyNextByte(int codeLength, int value, int index) {
-    return value << ((codeLength - index) << 3);
+    return (long) value << ((codeLength - index) << 3);
   }
 
   private static long applyType(long code, int codeLength, Type type) {
